Switch `PROJECTILE` to use `PagedEntityContainer<PROJECTILE>` as backing storage
The game still uses `psProjectileList` to maintain stable and
predictable order of iteration for projectiles, but the
individual `PROJECTILE` instances are allocated from the
global `PagedEntityContainer<PROJECTILE>` instance.

`proj_SendProjectileAngled()` now doesn't automatically
add the spawned penetrating projectile to `psProjectileList`,
which is a little step back in terms of ease of use (need
to additional code so that spawned projectiles are added
to the `psProjectileList` manually in relevant cases).

But it allows to remove additional checks inside the
tight loop in `proj_UpdateAll()`, which enumerates all
projectiles currently in play, thus considerably speeding
up this frequent and expensive operation.

Signed-off-by: Pavel Solodovnikov <pavel.al.solodovnikov@gmail.com>