pathfinding: Introduce path heatmap for improving crowd-avoidance behavior
The patch includes `PathHeatmap` class, which is a singleton with
the ability to create read-only snapshots via calls `takeSnapshot()` method.

The current solution to collision avoidance is to deposit heat right
where all the droids are at a given point in time.
This heat information is then accounted for in A* pathfinding code (namely,
in `fpathNewNode()`), so that scaled heat values are added the cost
of candidate nodes for OPEN set of A* algorithm.

Internal structure uses a rolling offset to avoid having to clear
and re-assign all tiles of the heatmap every tick. The offset is advanced
once per simulation tick by calling `advanceOffset()`, which naturally
support "heat dissipation" over time without the need to explicitly update
every tile on each simulation tick.

NOTE: internal heat values are represented by an 32-bit integer per tile.
There is also an owner ID per tile, which helps to avoid self-interference
when a droid queries the heat value of a tile it has just added heat to.

WARNING: this class is NOT thread-safe. It is expected that only
the main thread will modify it, while pathfinding threads will only
read from read-only snapshots taken at the start of each simulation
tick (created by calling `takeSnapshot()`).

CPU-based debug overlay is also included, which may be enabled with
`LCTRL + N` combination.

Signed-off-by: Pavel Solodovnikov <pavel.al.solodovnikov@gmail.com>